main/adding in images for main and gameover
These screens are the gameover screen, the main menu/title screen, as
well as the images that will act as buttons when pressed on interacted
with. I am planning on creating a third enemy soon and making the
enemies random afterwards with it switching between a lady in the purple
dress, butler, and the third enemy.